Restructure website folders
