+++ services/marko-example-website/app/routes/content.js
@@ -1,0 +1,8 @@
+const { withContent } = require('@base-cms/marko-web/middleware');
+const content = require('../templates/content');
+
+module.exports = (app) => {
+  app.get('/:prefix(*):id(\\d{8}):suffix(*)', withContent({
+    template: content,
+  }));
+};
