+++ services/marko-example-website/app/routes/dynamic-page.js
@@ -1,0 +1,8 @@
+const { withDynamicPage } = require('@base-cms/marko-web/middleware');
+const page = require('../templates/dynamic-page');
+
+module.exports = (app) => {
+  app.get('/page/:alias', withDynamicPage({
+    template: page,
+  }));
+};
