+++ services/marko-example-website/app/templates/content/index.marko
@@ -1,6 +1,6 @@
 import { get } from '@base-cms/object-path';
 
-<site-layout>
+<website>
   <site-navbar />
   <div class="container-fluid">
     <h1>${get(data, "content.name")}</h1>
@@ -13,4 +13,4 @@ import { get } from '@base-cms/object-path';
   </div>
   <hr>
   <site-footer />
-</site-layout>
+</website>
