+++ services/marko-example-website/app/templates/index.marko
@@ -1,6 +1,6 @@
-<site-layout>
+<website>
   <site-navbar />
   <h1>Homepage</h1>
   <hr>
   <site-footer />
-</site-layout>
+</website>
