--- services/marko-example-website/app/templates/section/tactical.marko
@@ -1,13 +1,0 @@
-import { get } from '@base-cms/object-path';
-
-<site-layout>
-  <@head>
-    <title>${get(data, "section.name")}</title>
-  </@head>
-
-  <site-navbar />
-
-  <div class="container-fluid">
-    <h1>Override! ${get(data, "section.name")}</h1>
-  </div>
-</site-layout>
